Move theme to design module
