+++ app/src/main/kotlin/uk/govuk/app/ui/GovUkApp.kt
@@ -41,6 +41,7 @@ import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.launch
 import uk.govuk.app.R
+import uk.govuk.app.design.ui.theme.GovUkTheme
 import uk.govuk.app.home.ui.navigation.homeGraph
 import uk.govuk.app.launch.AppLaunchState
 import uk.govuk.app.launch.AppLaunchViewModel
@@ -48,7 +49,6 @@ import uk.govuk.app.navigation.TopLevelDestination
 import uk.govuk.app.onboarding.ui.navigation.ONBOARDING_GRAPH_ROUTE
 import uk.govuk.app.onboarding.ui.navigation.onboardingGraph
 import uk.govuk.app.settings.ui.navigation.settingsGraph
-import uk.govuk.app.ui.theme.GovUkTheme
 
 private const val SPLASH_ROUTE = "splash"
 private const val ONBOARDING_COMPLETED_ROUTE = "onboarding_completed"
@@ -104,7 +104,7 @@ private fun SplashScreen(
         modifier = Modifier
             .fillMaxHeight()
             .fillMaxWidth()
-            .background(GovUkTheme.colourScheme.surfaces.backgroundSplash),
+            .background(GovUkTheme.colourScheme.surfaces.primary),
         verticalArrangement = Arrangement.Center
     ) {
         val composition by rememberLottieComposition(
